requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.10
langfuse>=2.16.1
langfuse-langchain>=2.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0
pytest 